requirements.txt
collections
networkx
pandas
dash
plotly.graph_objects